Call badResult in services

+++ habitcircle/server/application/services/circle.service.ts
@@ -7,7 +7,7 @@ import { Habit } from "@/server/domain/entities/habit.entity";
 import { DeleteCircleCommand, DeleteCircleResult } from "../dto/circle/delete-circle.dto";
 import { AddHabitCommand, AddHabitResult } from "../dto/circle/add-habit.dto";
 import { HabitRepository } from "@/server/domain/repositories/habit.repository";
-import { UNEXPECTED_ERROR } from "@/lib/constants";
+import { badResult } from "@/lib/utils";
 
 export class CircleService {
     constructor(
@@ -20,7 +20,7 @@ export class CircleService {
         try {
             const owner = await this.userRepo.findById(cmd.ownerId);
             const members = await Promise.all(
-                cmd.memberIds.map(this.userRepo.findById)
+                cmd.memberIds.map(id => this.userRepo.findById(id))
             );
             const circle = Circle.create(
                 cmd.circleName,
@@ -36,8 +36,7 @@ export class CircleService {
             return { ok: true, value: { circleId: circle.id, name: circle.name.get() } }
 
         } catch (err) {
-            const message = err instanceof Error ? err.message : UNEXPECTED_ERROR;
-            return { ok: false, error: message }
+            return badResult(err);
         }
     }
 
@@ -50,23 +49,28 @@ export class CircleService {
             const circle = await this.circleRepo.findById(cmd.circleId);
             const circleWithHabit = circle.addHabit(habit);
             
-            await this.habitRepo.save(habit);
-            await this.circleRepo.save(circleWithHabit);
+            await Promise.all([
+                this.habitRepo.save(habit),
+                this.circleRepo.save(circleWithHabit)
+            ])
             
             return { ok: true, value: { result: true } }
         } catch (err) {
-            const message = err instanceof Error ? err.message : UNEXPECTED_ERROR;
-            return { ok: false, error: message };
+            return badResult(err);
         }
     }
 
     async delete(cmd: DeleteCircleCommand): Promise<Result<DeleteCircleResult>> {
+        const circle = await this.circleRepo.findById(cmd.circleId);
+        const circleOwnerId = circle.getOwner().id;
+        if (circleOwnerId !== cmd.requestingUserId)
+            return badResult("Cannot delete circle you do not own");
+
         try {
             await this.circleRepo.delete(cmd.circleId)
             return { ok: true, value: { result: true } }
         } catch (err) {
-            const message = err instanceof Error ? err.message : UNEXPECTED_ERROR;
-            return { ok: false, error: message }
+            return badResult(err);
         }
     }
 }
